pyproject.toml
[project]
name = "superfinance-mcp"
version = "0.1.0"
description = "SuperFinance MCP server - Advanced financial data and analysis tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "yfinance>=0.2.62",
    "uvicorn>=0.30.0",
    "snaptrade-python-sdk>=11.0.154",
    "upstash-redis>=1.0.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "financedatabase>=2.3.1",
    "financetoolkit>=2.0.6",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "black>=24.2.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.bandit]
exclude_dirs = ["tests"]
test = ["B101", "B102", "B103", "B104", "B105", "B106", "B107", "B108", "B110", "B112", "B201", "B301", "B302", "B303", "B304", "B305", "B306", "B307", "B308", "B309", "B310", "B311", "B312", "B313", "B314", "B315", "B316", "B317", "B318", "B319", "B320", "B321", "B323", "B324", "B325", "B401", "B402", "B403", "B404", "B405", "B406", "B407", "B408", "B409", "B410", "B411", "B412", "B413", "B501", "B502", "B503", "B504", "B505", "B506", "B507", "B508", "B509", "B601", "B602", "B603", "B604", "B605", "B606", "B607", "B608", "B609", "B610", "B611", "B612", "B613", "B614", "B615", "B616", "B617", "B618", "B619", "B620", "B621", "B622", "B623", "B624", "B625", "B626", "B627", "B628", "B629", "B630", "B631", "B632", "B633", "B634", "B635", "B636", "B637", "B638", "B639", "B640", "B641", "B642", "B643", "B644", "B645", "B646", "B647", "B648", "B649", "B650", "B651", "B652", "B653", "B654", "B655", "B656", "B657", "B658", "B659", "B660", "B661", "B662", "B663", "B664", "B665", "B666", "B667", "B668", "B669", "B670", "B671", "B672", "B673", "B674", "B675", "B676", "B677", "B678", "B679", "B680", "B681", "B682", "B683", "B684", "B685", "B686", "B687", "B688", "B689", "B690", "B691", "B692", "B693", "B694", "B695", "B696", "B697", "B698", "B699", "B700", "B701", "B702", "B703", "B704", "B705", "B706", "B707", "B708", "B709", "B710", "B711", "B712", "B713", "B714", "B715", "B716", "B717", "B718", "B719", "B720", "B721", "B722", "B723", "B724", "B725", "B726", "B727", "B728", "B729", "B730", "B731", "B732", "B733", "B734", "B735", "B736", "B737", "B738", "B739", "B740", "B741", "B742", "B743", "B744", "B745", "B746", "B747", "B748", "B749", "B750", "B751", "B752", "B753", "B754", "B755", "B756", "B757", "B758", "B759", "B760", "B761", "B762", "B763", "B764", "B765", "B766", "B767", "B768", "B769", "B770", "B771", "B772", "B773", "B774", "B775", "B776", "B777", "B778", "B779", "B780", "B781", "B782", "B783", "B784", "B785", "B786", "B787", "B788", "B789", "B790", "B791", "B792", "B793", "B794", "B795", "B796", "B797", "B798", "B799", "B800", "B801", "B802", "B803", "B804", "B805", "B806", "B807", "B808", "B809", "B810", "B811", "B812", "B813", "B814", "B815", "B816", "B817", "B818", "B819", "B820", "B821", "B822", "B823", "B824", "B825", "B826", "B827", "B828", "B829", "B830", "B831", "B832", "B833", "B834", "B835", "B836", "B837", "B838", "B839", "B840", "B841", "B842", "B843", "B844", "B845", "B846", "B847", "B848", "B849", "B850", "B851", "B852", "B853", "B854", "B855", "B856", "B857", "B858", "B859", "B860", "B861", "B862", "B863", "B864", "B865", "B866", "B867", "B868", "B869", "B870", "B871", "B872", "B873", "B874", "B875", "B876", "B877", "B878", "B879", "B880", "B881", "B882", "B883", "B884", "B885", "B886", "B887", "B888", "B889", "B890", "B891", "B892", "B893", "B894", "B895", "B896", "B897", "B898", "B899", "B900", "B901", "B902", "B903", "B904", "B905", "B906", "B907", "B908", "B909", "B910", "B911", "B912", "B913", "B914", "B915", "B916", "B917", "B918", "B919", "B920", "B921", "B922", "B923", "B924", "B925", "B926", "B927", "B928", "B929", "B930", "B931", "B932", "B933", "B934", "B935", "B936", "B937", "B938", "B939", "B940", "B941", "B942", "B943", "B944", "B945", "B946", "B947", "B948", "B949", "B950", "B951", "B952", "B953", "B954", "B955", "B956", "B957", "B958", "B959", "B960", "B961", "B962", "B963", "B964", "B965", "B966", "B967", "B968", "B969", "B970", "B971", "B972", "B973", "B974", "B975", "B976", "B977", "B978", "B979", "B980", "B981", "B982", "B983", "B984", "B985", "B986", "B987", "B988", "B989", "B990", "B991", "B992", "B993", "B994", "B995", "B996", "B997", "B998", "B999"]
